UserConnection: do not inherit from shared_from_this + size_t->uint conversion warning fix.

+++ src/Plugins/TundraLogic/TundraLogicUtils.cpp
@@ -18,7 +18,7 @@ std::vector<s8> StringToBuffer(const String& str)
 String BufferToString(const std::vector<s8>& buffer)
 {
     if (!buffer.empty())
-        return String((const char*)&buffer[0], buffer.size());
+        return String((const char*)&buffer[0], static_cast<uint>(buffer.size()));
     else
         return String();
 }
